Split files to improve organization
